+++ lib/e_menu/src/pkgIndex.tcl
@@ -1,4 +1,4 @@
-package ifneeded apave 4.1.4 [list source [file join $dir apave.tcl]]
+package ifneeded apave 4.1.5 [list source [file join $dir apave.tcl]]
 
 # A short intro (for Ruff! docs generator:)